pyproject.toml
[project]
name = "kv4006-render"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.4",
    "livereload>=2.7.0",
    "pyyaml>=6.0.2",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "nuitka>=2.5.6",
]
